drc,nvmain: Simplify capacity calculations in LO-Cache and base MemoryController
The base memory controller uses the device size, bus width, and number of beats
to calculate the simulated capacity of a memory channel. In 3D-stacked memories
we assume devices will not exist, so the calculation is simplified to use the
memory word size, defined by the bus width and number of beats only.

--HG--
branch : cleanup