gem5: Fix potential deadlock scenario with large memory sizes
Most gem5 ISAs do not use contiguous memory regions, while NVMain does. This
can cause deadlocks when non-contiguous regions (i.e., non masterInstance) in
the gem5 interface are woken up after switching from atomic to timing mode.
This patch changes event scheduling to always occur on the masterInstance
such that it is not possible to schedule an event which will be ignored by
the interfaces tick method, causing a deadlock in gem5.